pyproject.toml
[tool.poetry]
name = "coopfinder"
version = "0.1.0"
description = ""
authors = ["Bryce Jenkins <bryce@brycejenkins.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-steam-api = "^2.0.2"
python-dotenv = "^1.0.1"
rich = "^13.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
